requirements.txt
fastapi~=0.112.1
uvicorn~=0.15.0
python-dotenv~=0.19.0
pyodbc~=4.0.32
